added swagger examples for Auth, Files, Finance, FinishedGoods, Inventory, Materials

+++ src/MyFactory.WebApi/Contracts/FinishedGoods/FinishedGoodsInventoryResponse.cs
@@ -3,6 +3,6 @@
 public record FinishedGoodsInventoryResponse(
     Guid SpecificationId,
     Guid WarehouseId,
-    int Quantity,
+    double Quantity,
     decimal UnitCost
 );
